Extract CategorizationMatcher collaborator
Pulls the read-only matching surface — pattern matching, exemplar lookup,
rule evaluation, and the uncategorized-transaction scan — into a new
categorization/matcher.py with a CategorizationMatcher class. Module-level
helpers _match_shape_label, _match_exemplar, and _match_text stay private
to matcher.py; match_merchants drops its leading underscore because the
facade calls it directly from apply_merchant_categories and
_categorize_items_inner.

CategorizationService now constructs self._matcher = CategorizationMatcher(db)
in __init__ and delegates match_merchant, fetch_active_rules,
match_first_rule, and find_matching_rule as thin pass-throughs to preserve
the public surface. Internal callers (apply_merchant_categories,
_categorize_items_inner, apply_rules, categorize_pending) use self._matcher
instead of the removed module-level helpers; the private
_fetch_uncategorized_rows method is gone in favor of
self._matcher.fetch_uncategorized_rows().

No behavior change; 2058 tests pass.
